# Fast integer scale unpooling