I added a code path for faster unpooling when the output shape is an integer multiple of the input shape. This happens e.g. when upsampling the previous result of 2x2 max pooling, which is a common use case. This PR speeds up the forward pass about 1.7-3.2 times for this case.

MWE:
With batch and channel size 1, this PR makes the unpooling 1.7 times faster:
```
# Prepare input data
In [1]: import numpy as np; import cupy; import chainer.functions as F; x = cupy.arange(1, 128*64+1).reshape((1, 1, 64, 128)).astype(np.float32)

# Without this PR (master branch)
In [2]: %timeit F.unpooling_2d(x, 2, outsize=(x.shape[2]*2, x.shape[3]*2))
178 µs ± 18.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

# With this PR
In [3]: %timeit F.unpooling_2d(x, 2, outsize=(x.shape[2]*2, x.shape[3]*2))
99.2 µs ± 93.2 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

With larger batch and channel size, this PR makes the unpooling 3.2 times faster:
```
# Prepare input data
In [1]: import numpy as np; import cupy; import chainer.functions as F; x = cupy.arange(1, 128*32*128*64+1).reshape((128, 32, 64, 128)).astype(np.float32)

# Without this PR (master branch)
In [2]: %timeit F.unpooling_2d(x, 2, outsize=(x.shape[2]*2, x.shape[3]*2))
20.1 ms ± 1.82 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

# With this PR
In [3]: %timeit F.unpooling_2d(x, 2, outsize=(x.shape[2]*2, x.shape[3]*2))
6.17 ms ± 17.8 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```