@tkerola 2019-05-13
@mitmul Thanks for the review! I fixed the import exception and commented about non-zero padding. Can you please check it?