@mitmul 2019-05-13
All the tests on the ChainerCI and examples/docs tests on Jenkins have passed. LGTM