@tkerola 2019-05-13
Hmm, it seems like the test for group normalization failed - unrelated to this PR. Can you please run the tests again?