@tkerola 2019-05-13
Jenkins, test this please