@mitmul 2019-05-13
@tkerola This PR passed all the tests. Please only see the ImosCI test jobs basically and only for the example and docs tests, please check the Jenkins tests. Thanks. (FYI, "Jenkins, test this please" triggers ImosCI and Jenkins at the same time)